Adapt to new liblouis API (typeform is now short instead of byte)

+++ liblouis-utils/liblouis-core/src/test/java/LiblouisCoreTest.java
@@ -133,7 +133,7 @@ public class LiblouisCoreTest extends AbstractTest {
 		assertArrayEquals(new String[]{"⠋⠕⠕⠃⠁⠗"},
 		                  provider.withContext(messageBus)
 		                          .get(query("(table:'foobar.cti')")).iterator().next()
-		                          .fromTypeformedTextToBraille().transform(new String[]{"foobar"}, new byte[]{Typeform.ITALIC}));
+		                          .fromTypeformedTextToBraille().transform(new String[]{"foobar"}, new short[]{Typeform.ITALIC}));
 	}
 	
 	@Test
